+++ packages/extension/example/string.dart
@@ -2,6 +2,6 @@ import 'package:extension/string.dart';
 
 main() {
   final capitalized =
-      capitalizeFirstLetter('capitalize my first letter, please!');
+      'capitalize my first letter, please!'.capitalizeFirstLetter();
   assert(capitalized == 'Capitalize my first letter, please!');
 }