+++ packages/extension/lib/extension.dart
@@ -1,2 +1,3 @@
-export './enum.dart';
-export './string.dart';
+export 'date.dart';
+export 'enum.dart';
+export 'string.dart';