+++ packages/extension/lib/string.dart
@@ -1,9 +1,40 @@
-/// Returns string with capitalized first letter
-///
-/// Example:
-/// ```dart
-/// assert(capitalizeFirstLetter('test), 'Test');
-/// ```
-String capitalizeFirstLetter(String string) => (string?.isNotEmpty ?? false)
-    ? '${string[0].toUpperCase()}${string.substring(1)}'
-    : string;
+final RegExp _emailRegularExpression = RegExp(
+    r"^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$");
+
+extension StringUtils on String {
+  /// Returns string with capitalized first letter
+  ///
+  /// Example:
+  /// ```dart
+  /// assert('test'.capitalizeFirstLetter(), 'Test');
+  /// ```
+  String capitalizeFirstLetter() => (this?.isNotEmpty ?? false)
+      ? '${this[0].toUpperCase()}${this.substring(1)}'
+      : this;
+
+  /// Returns true if string contains email
+  bool get isEmail => _emailRegularExpression.hasMatch(this.toLowerCase());
+}
+
+/// Plural forms for russian words
+/// [number] count quantity for word
+String plural(int number, String form1, String form2, String form3) {
+  final num = number % 100;
+
+  if (num >= 11 && num <= 19) {
+    return form3;
+  }
+
+  final i = num % 10;
+
+  switch (i) {
+    case 1:
+      return form1;
+    case 2:
+    case 3:
+    case 4:
+      return form2;
+    default:
+      return form3;
+  }
+}